software-map: move snapshot and diff-summary helpers to software-map-snapshot
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
